+++ swig/Source/DOH/fio.c
@@ -298,9 +298,9 @@ int DohvPrintf(DOH *so, const char *format, va_list ap) {
 	  }
 	  if (strlen(encoder)) {
 	    enc = encode(encoder, Sval);
-	    maxwidth = maxwidth + strlen(newformat) + Len(enc);
+	    maxwidth = maxwidth + (int) strlen(newformat) + Len(enc);
 	  } else {
-	    maxwidth = maxwidth + strlen(newformat) + Len(Sval);
+	    maxwidth = maxwidth + (int) strlen(newformat) + Len(Sval);
 	  }
 	  *(fmt++) = 's';
 	  *fmt = 0;
@@ -314,7 +314,7 @@ int DohvPrintf(DOH *so, const char *format, va_list ap) {
 	  } else {
 	    nbytes += sprintf(stemp, newformat, Data(Sval));
 	  }
-	  if (Writen(so, stemp, strlen(stemp)) < 0)
+	  if (Writen(so, stemp, (int) strlen(stemp)) < 0)
 	    return -1;
 	  if ((DOH *) Sval != doh) {
 	    Delete(Sval);
@@ -340,7 +340,7 @@ int DohvPrintf(DOH *so, const char *format, va_list ap) {
 	  } else {
 	    enc = 0;
 	  }
-	  maxwidth = maxwidth + strlen(newformat) + strlen((char *) doh);
+	  maxwidth = maxwidth + (int) strlen(newformat) + (int) strlen((char *) doh);
 	  *(fmt++) = 's';
 	  *fmt = 0;
 	  if ((maxwidth + 1) < OBUFLEN) {
@@ -349,7 +349,7 @@ int DohvPrintf(DOH *so, const char *format, va_list ap) {
 	    stemp = (char *) DohMalloc(maxwidth + 1);
 	  }
 	  nbytes += sprintf(stemp, newformat, doh);
-	  if (Writen(so, stemp, strlen(stemp)) < 0)
+	  if (Writen(so, stemp, (int) strlen(stemp)) < 0)
 	    return -1;
 	  if (stemp != obuffer) {
 	    DohFree(stemp);
@@ -360,7 +360,7 @@ int DohvPrintf(DOH *so, const char *format, va_list ap) {
       } else {
 	*(fmt++) = *p;
 	*fmt = 0;
-	maxwidth = maxwidth + strlen(newformat) + 64;
+	maxwidth = maxwidth + (int) strlen(newformat) + 64;
 
 	/* Only allocate a buffer if it is too big to fit.  Shouldn't have to do
 	   this very often */
@@ -395,7 +395,7 @@ int DohvPrintf(DOH *so, const char *format, va_list ap) {
 	default:
 	  break;
 	}
-	if (Writen(so, stemp, strlen(stemp)) < 0)
+	if (Writen(so, stemp, (int) strlen(stemp)) < 0)
 	  return -1;
 	if (stemp != obuffer)
 	  DohFree(stemp);
@@ -408,7 +408,7 @@ int DohvPrintf(DOH *so, const char *format, va_list ap) {
   if (state) {
     int r;
     *fmt = 0;
-    r = Writen(so, fmt, strlen(fmt));
+    r = Writen(so, fmt, (int) strlen(fmt));
     if (r < 0)
       return -1;
     nbytes += r;
@@ -449,7 +449,7 @@ int DohPrintv(DOHFile * f, ...) {
     if (DohCheck(obj)) {
       ret += DohDump(obj, f);
     } else {
-      ret += DohWrite(f, obj, strlen((char *) obj));
+      ret += DohWrite(f, obj, (int) strlen((char *) obj));
     }
   }
   va_end(ap);
